LAr Calorimeter Signal FeedThroughs detailed geometry implemented.
LAr Calo DeadMaterial description was simplified for the Barrel and EC FeedThroughs and Front-End Crates.
The introduced changes are:
1) Detailed description for each Signal FT, including corresponding materials.
2) Extra structures for Endcap FTs (cables, chimneys).
3) Endcap FEC size and position fixes.

To activate, one have to set
 from LArGeoAlgsNV.LArGeoAlgsNVConf import LArDetectorToolNV
 LArDetectorToolNV.ActivateFeedThrougs=True
Otherwise, the earlier behavior remains.

The patch contains slight change of CALO and Muon envelope volumes
(a 12 mm Z plane shift at the back of the calorimeter).