Manually sweeping DerivationFrameworkInDet/InDetTrackSelectionToolWrapper from 21.2 to master
The tool InDetTrackSelectionToolWrapper is required to allow the derivation framework to decorate InDetTrackParticles with track selection flags and thereby allow their selection. This MR sweeps the tool from 21.2 to master. In consequence of this MR the full set of InDetTrackParticle thinning criteria can be applied in DAOD_PHYS; this is also set in the MR.