Fix the ID Trigger naming convention in the ID Trigger configuration code, add test beamspot chain
There were inconsistencies in the way the ID Trigger sequences were configured.
These changes partly address thie to tidy the configuration code somewhat.
As part of this is a fix to the naming convention in the beamspot chains, together
with the implementation of a test beamspot chain in the LS2 menu. This includes a
renaming of the beamspot chain in the pp_run3 menu, to conform to the established
naming conventin from Run 2, that has been agreed with Menu coordination.