Add Bkg_TruthEvent to the list of known McEventCollection names in HepMcParticleLink.cxx
The pre-mixing jobs rename the output `McEventCollection` to `Bkg_TruthEvent` rather than `TruthEvent`.
`HepMcParticleLink` doesn't know that this is a possible collection name that it should check for so
all links are broken and the attempts to find the `McEventCollection` in the T->P converter methods
always return `nullptr`. This causes the above crash as there is currently no protection against
dereferencing a null `McEventCollection` pointer. Adding `Bkg_TruthEvent` to the list of known
`McEventCollection` names in `HepMcParticleLink.cxx` should fix ATLASSIM-4539.