Add a job option to allow the Bjet vertex tracking to run an actual fullscan rather than the superRoi tracking
Just adds a flag which modifies the creation of the RoiDescriptor in the superRoi
building code.

This is all it needs to switch between fullscan and superRoi processing - changes for
this in the configuration should be simply straightforward - either set this single
job option, or don't, it really should not be any more complicated that that.