Porting DerivationFrameworkMCTruth to release 22: sweep of python directory
Thus begins what is likely to be a fairly involved process to port the MC truth machinery to master. This MR is a simple sweep of the DerivationFrameworkMCTruth/python directory from 21.2, since the python hasn't evolved in r22 at all. This MR will not do anything in master for now, since these python modules are only used in the derivation framework. It is a pre-requisite for getting the truth content into the derivations in release 22.