MGC updates for use of the central PMG parameters
This adds several functions to the param helpers to:
 - Get a list of masses from a param card
 - Get a list of widths from a param card
 - Get the value from an arbitrary block in a param card
 - Get the centrally provided PMG params for the mass, width, and yukawa
blocks
 - Test if the values match those centrally provided values
 - Set the param card to match those centrally provided values

Within the main utils functions, the new_process and generate functions
have been updated to allow users to request the automatic setting of
parameters based on the PMG defaults, and to raise an exception if the
parameters do not match the PMG defaults. For the time being, both
pieces of functionality are disabled, since we don't want to change
defaults under people or suddenly have crashing jobs without the
blessing of PMG. This should take care of !31896

Thanks again to Silke Moebius, who provided the first draft of this code
(in that MR) and implemented the underlying infrastructure that makes
this possible!