Updated the xAODCaloEvent package's dictionary generation.
Made sure that all potentially useful/necessary types are generated
in an organised fashion, using the helper code from xAODCore.

Removed code not necessary with modern ROOT versions.

Declared the typedef names to the generated ROOT dictionary.

Moved the pure STL types to the xAODCoreSTLDict dictionary. At the
same time fixed the comment in xAODCoreSTLDict.h about the meaning
of "3" in std::bitset<3>.