CaloUtils: Add ToolWithConstants.
ToolWithConstants is a replacement for ToolWithConstantsMixin that should
work with Gaudi's new property handling as well as MT conditions.

The constants are now found from the conditions store.  This implies
that the mode of reading a pool file directly, without cool,
is no longer supported.  (That was only used for testing anyway.)
This code also does not directly handle inline cool data; the plan is to handle
that via a separate conditions algorithm which will build ToolConstants
objects from inline data.  Accordingly, the property giving the
conditions key is now called DBHandleKey, rather than detStoreKey.

This is now not a mixin class, instead being templated on the base class,
to be more consistent with other Gaudi code.