Use auto-declaring property ctor for BCM RDOs ReadHandleKey
Fatras does not simulate the BCM, and therefore when
RawInfoSummaryForTagWriter attempts to retrieve BCM
RDOs, dereferencing the associated VarHandle fails
and leads to a crash.

Instead of introducing a new property explicitly, this
commit uses the auto-declaring property constructor
for the associated ReadHandleKey;
to disable the reading of BCM RDOs, this key can be made
empty in job configuration, which will record zero BCM hits
in the RISFTobject.