Fix McParticleEvent TruthParticle_test unit test
After the updates in !32855 it is necessary to record the `McEventCollection` being linked to in
StoreGate. Part of the fix for ATLASSIM-4586.

Also patched the `TruthParticleContainer`, so that it is explicit that the position of `GenEvent` objects
within the `McEventCollection` rather than the `GenEvent::event_number` is being used to create
`HepMcParticleLink` objects.