MC truth in release 22 verison of DAOD_PHYS
This MR makes use of !33167 and !33139 and introduces MC truth into DAOD_PHYS in master. Necessarily a number of components have been commented out as they rely on other software that has not been migrated into release 22 and will need to be added in further MRs. The missing components are as follows. If no file name is provided the commented lines are in MCTruthCommon.py.
- Q/G decoration jets (TruthDerivationTools.py)
- TruthMetaDataWriter
- TruthCategories (PHYS.py)
- CopyTruthJetParticles
- addAntiKt4TruthDressedWZJets
- truth_groomed
- scheduleTauTruthTools
- GenFilterTools in GenFilterToolSetup.py
- addTruthEnergyDensity
- TauTruth (PHYS.py)
- LHE3 weights (PHYS.py)
- addTruthJets (PHYS.py) - seg fault if uncommented
- TruthQGDecorationTool
- TauTruth omitted from TruthNavigationDecorator
- addLargeRJetD2 (complains about missing AntiKt10...R20Jets