InDetEventCnvTools: MT-friendly version of prepareRIO_OnTrack.
IEventCnvSuperTool::prepareRIO_OnTrack was used to fix up PRD ElementLinks
before writing.  However, this was done on objects in place, violating
const-correctness.  This is not MT-safe.

Add a new interface prepareRIO_OnTrackLink.  This takes the object
as a const pointer and instead of modifying it, returns the new
ElementLink components.  The T->P converter can then use this to write.