PyUtils: Improve startup time of acmd.py
In most cases it is not necessary to load all plugins as the user very
likely specified one on the command line. Try to first load a specific
plugin and then fall back to loading all of them. This reduces the
startup time significantly, i.e. if we keep adding new sub-commands.

Also remove the "running sub-command" INFO message as this makes it
impossible to implement sub-commands that can be used in pipes.