Partial support for TruthTaus in DAODs
Hello,

This MR imports from 21.2 the functionality to build truth taus when running on xAODs.
At the moment, the production of TruthTaus from EVNT does not work, because for that we'll need to migrate DerivationFrameworkMCTruth/src/TruthCollectionMakerTau.cxx.
The TruthTaus collection is now enabled in MCTruthCommon.py, and in DAOD_PHYS as well.
I took this opportunity to import the smart slimming lists for DiTauJetsCP and DiTauJetsLowPt.
Some functionality in TauCommon has also been imported from 21.2.
The on-the-fly scheduling of DiTauJetsLowPt does not work yet, due to a missing dependency in DerivationFrameworkJetEtMiss/python/ExtendedJetCommon.py (addCHSPFlowObjects).
Not sure if this workflow is supposed to be different in R22.

Naive remark: the truth tau building has not been extensively validated. To come: the comparison of TruthTaus built from EVNT, in R21.2 and R22, when we'll have TruthCollectionMakerTau (the comparison of TruthTaus on xAOD is more cumbersome).

Cheers,
Bertrand