G4AtlasTools: Fix placement of using declarations.
Do not put a using declaration in the global namespace before an #include.

Unfortunately, it appears that the usage here was to work around
broken geant headers that expect CLHEP names in the global namespace.
To avoid warnings from the namespace checker, change the using
declarations of constants to declarations of new (static, constexpr)
constants.