Re-enable MDT decoding in new JO tests
Re-enabling the MDT decoding for the new JO test:
- Commented out the MDT region selector tool definition, as this needs proper configuration for the new JO test. This introduces a warning for the MdtRdoToPrepData decoding, since the region selector tool is not defined (this is why the reference files for the decoding test is updated). However, without running the decoding at all, we can not run the trigger algorithmd, so is a slight improvement.
- This fixes the errors reported in ATR-21280

Also moved the decoding out of the L2SA job options. It needed to be updated to run inside the event views (otherwise we run into problems running the decoding multiple times in the full event when there is more than one RoI), and is not specific to L2SA, so was moved to the muon menu setup for now.