CaloUtils: Add fast no-context retrieval in ToolWithConstants.
The generic way to retrieve a constant in ToolWithConstants is to call
operator() on it, passing in a Context object.  Added another operator()
which does not take arguments.  This can be used only if the constant
has been initialized from job options.  (The intended use case is for
the L2 trigger.)