egammaTools: ToolWithConstants work.
An issue with the cluster correction metatool is that data dependencies
don't get propagated through it.

With the transition to ToolWithConstants, the metatool is disabled.
That means that data dependencies _are_ propagated from the correction
tools to the parent.

However, egammaSwTool is creating correction tools for all possible
cluster sizes, both for regular and supercell clusters.

This means that all jobs see dependencies in the calo cells,
even trigger jobs which use only supercell clusters and don't
make the normal calo cells.

It's actaully wasteful for this tool to build correction tools for
both normal and supercell clusters, since both can't be used at the
same time.

For now, add options to configure an egammaSwTool for one or the other
cases.