CaloClusterCorrection: Switch cluster corrections to ToolWithConstants.
Switch the cluster correction classes from ToolWithConstantsMixin
to ToolWithConstants.

The metatool (CaloRunClusterCorrections) does not yet work with the new
class, so it has been disabled.   This implies that one cannot configure
the list of corrections from only a cool tag.  In many cases, though,
the cool tag had already been replaced by a static version.
For now, adjust the configuration so that we always use a static
configuration, rather than the tag.  This should be revisited later.

The configuration scripts also need some further cleanup, to be done
in a future change.