Remove defunct job option from ByteStreamInputSvc
The ByteStreamEventStorageInput service provided a legacy job option called FullFileName to specify bytestream input. With the move to EventSelector this option was all but removed from ByteStreamEventInputSvc. The EventSelectorByteStream provieded the input option to specify input. It captured the FullFileName option of the InputSvc. This commit removes the relevant parts of the ByteStreamEventStorageInputSvc and EventSelectorByteStream. It also attempts to move all references in other job options (unless they specify being for an older release).