AnalysisBase: Add dummy flake8 configuration
`flake8` is not available in AnalysisBase. While our cmake configuration
does handle this correctly, it fails once developers start using special
flake8 options, e.g.:
```
... POST_BUILD_CMD ${ATLAS_FLAKE8} --enable-extensions ATL902
```
Even if `ATLAS_FLAKE8` is not defined it will now try to run a command
of name `--enable-extensions`, which of course fails. An easy workaround
is to `set( ATLAS_FLAKE8 true )` where `true` here is the shell builtin,
which returns success and swallows any command line args given to it.