Minor updates for NSW decoding in the trigger
Just making the instances of RDOtoPRD decoding algorithms for the NSW detectors distinct so we don't run into interference issues in the future (probably the only reason it works OK now is that the seeded decoding isn't yet implemented). Also slightly updating the naming for the MM decoding to be consistent with the other technologies (which makes searching the log files easier).