InDetTrigAmbiguitySolver: Complain about misconfigured MT jobs.
The thread-safety checker was not properly checking calls to virtual functions.
When that is fixed, we get a warning about the call to
IPRDtoTrackMapExchange::setPRDtoTrackMap.

This is indeed not thread safe, and is supposed to be used only for
backwards compatibility with run 2.

To resolve this, split out the non-thread-safe part into a separate function.
We suppress the thread-safety checker warning in the call to it,
but we also explicitly check and fail the job if we're running
such a run 2 configuration with more than one thread.