TriggerMenuMT: Stop producing L1 menu files during build
The main change is to no longer produce the L1 menu files (.xml, .json)
during the build and not install them any longer as part of the release.
This technique was inherited from `TriggerMenuXML`, but no current job
actually uses these pre-generated files. This simplifies the cmake
configuration greatly and `TriggerMenuMT` is no longer forced to run at
the end of the build.

To ensure menus are working, add a unit test for each L1 menu.

Some unrelated package cleanup:
- Delete unused `testMenuMT.sh`, `trigermenuMT_test.py`, `generateMenuMT.py`.
- Delete `generateMT.py` and instead inline the flags in `test_HLTmenu.sh`.
- Move `menu_config_tests.py` to `python/` directory.