RootUtils: Fix py38 memory corruption.
With python 3.8, the hacks used in pythonize_read_root_file lead to heap
corruption.  Rework to avoid this.

We're rather constrained since this has to work with both py2 and 3,
so the way it's done here is a bit roundabout and ineffecient.
Once we drop py2, this can be done much better by simply
having read_root_file return a PyBytes object