Fixing mistakes in the code.
I forgot to update the code used to execute tasks on the CPU in
AthCUDA::KernelRunnerSvc after changing the return types from
StatusCode to int. This was making the execution on CPU always
fail.

I also removed the unit test using the "standalone" CUDA example,
as that one can not run correctly when compiled with CUDA, but
executed on a host that has no GPUs in it. (Which is the setup that
we use on the CI build nodes.)