TrigPSC: Suppress thread-safety checker warning.
An uncoming change to the thread-safety checker will generate warnings
for virtual calls to unsafe functions.

This then gets a warning for the call to ITrigEventLoopMgr::prepareForRun
in Psc::prepareForRun.  Perhaps that latter function should also be marked
NOT_THREAD_SAFE, but that's awkward since the base class defining this
interface is in hltinterface.

Just suppress the warning for now.