DAOD_PHYS: add low-pt di-tau reconstruction
Hello,

This MR enables the on-the-fly reconstruction of low-pt di-taus in DAOD_PHYS.
A small change was made in the DiTauRec tool scheduling, to avoid a "spurious JetAlgorithm addition" to AlgSequence, which could cause some conflicts when scheduling ditau reconstruction on-the-fly, depending when the DAOD_PHYS AthSequencer is added to DerivationFrameworkJob.
Other small changes: the import of `DerivationFramework__CommonAugmentation` and `DerivationFramework__DerivationKernel` was commented out, it did not seem mandatory, please let me know if that is wrong.
Also, the `DerivationFramework__DerivationKernel` was added to `DerivationFrameworkJob` instead of `SeqPHYS`, in case that was intended, I am happy to put this back.

Besides, TauCommon has been better structured, so that functions can be imported and more specific actions can be made, such as decorate only the desired tau ID WP, instead of all 4 WPs.

What is not functional yet is the thinning of low-pt ditaus. No matter what I tried, I end up with `ERROR MultipleProxyLoader: unable to find valid proxy loader for DiTauJetsLowPt.nSubjets`.
Not sure why, the ditau EDM is defined in xAODTau, and that was working in 21.2. To be followed up.

Cheers,
Bertrand