Remove remaining const_casts from CombinedMuonTrackBuilder
Three of the four const_casts did not involve objects from SG: they were necessitated by the fact that CombinedMuonTrackBuilder used ITrackFitter interfaces, which assumed const tracks.
As it turns out, there is only one case in which the ITrackFitter interface to the CombinedMuonTrackBuilder was actually used, by the MuonSegmentRegionRecoveryTool.
The MuonSegmentRegionRecoveryTool uses the global chi2 fitter in EO-only mode, and the CombinedMuonTrackBuilder otherwise.
So a CombinedMuonTrackBuilder ToolHandle has been added to the MuonSegmentRegionRecoveryTool, and the configuration updated appropriately.
Depending on whether m_onlyEO is true or false, the appropriate tool is retrieved and used in the fit.
The interfaces in ICombinedMuonTrackBuilder are now non-const in cases where a const track is not expected, and the const_casts have been removed.