AthenaKernel: Change overlap handling of CondCont
We were allowing ranges in CondCont to overlap one another.
This isn't just academic, as we can frequently get such ranges
from IOVDbSvc.
However, this was leading in some cases to a lookup failure.

Change the semantics of overlap/extend so that we no longer overlap ranges.
Rather, in case of a overlap, the new range is truncated to avoid the overlap.
(We also allow treating an overlapping final range as an extension.)
Note that it is in principle possible for the truncated part
of the new range to contain the current IOV key.  The caller
will need to handle that case.