Update ISF_HitAnalysis.cxx: check StatusCode retrieving MuonEntry
If the 'MuonEntry' `TrackRecordCollection` is not found, the
algorithm crashes; I don't think this is the desired result.
Following the `ATH_CHECK` to retrieve the collection, the
`StatusCode` is checked but wasn't assigned for us in this call
to `SG::retrieve()`.

This commit replaces the `ATH_CHECK` in favor of an explicit check
of the return value of the `SG::retrieve()`. Now, if the
`TrackRecordCollection` is not found, a warning is displayed and
the algorithm continues.

Also updated the copyright.