LArRawConditions: Change std::pair selection patterns back to explicit names.
In the selection.xml files, we were using patterns to select std::pair
classes.  The reason is that we were earlier seeing errors if we did not
also have dictionaries for the __pair_base classes.

But these patterns are causing issues, either selecting too much 
or too little.  Further, it appears that with the current versions
of root and gcc, the pair_base dictionaries are no longer needed.
So switch the patterns back to explicit names.