Disable RingSetConf in-file metadata
At the moment the RingSetConf in-file metadata is not friently to
MT-processing. As it is accessed by an algorithm it would need migrating
to MetaCont and ReadMetaHandle. However it is also safe to configure the
algorithm to neither read or write the metadata content. This disables
writing the RingSetConf. When the developers wish to return to using it
they will also modernize their usage of the in-file metadata
infrastructure to make it either thread safe or MT-friendly.