Fix crash in finalisation of the standalone post processing
For some reason, when compiling with cmake, ths standalone job
crashes in the finalisation of a correlation plot.

The changes the type of fit that root is trying and so should
no longer crash.

Interestingly, when compiled outside athena, or with the athena
release, but with a standalone makefile, this runs fine.