Loading event-level condition data intofast trackfinder view setup
Changes from Ben to isaolte view from Event Context processing prevent the
fall through of the setup of the pixel and SCT conditions because of the field
service. This should hopwfully get the service set up and fix issues seen in
ATR-21685. Hopefully nothing should be needed for the SCT or TRT.