Sweeping 21.2 version of SGxAODProxyLoader into master
The version of SGxAODProxyLoader in master, which defines the type for each container in the ExpressionEvaluation, has not been touched for four years. In consequence it misses several containers that were added later (particularly the DiTauJets) This MR sweeps the version from 21.2 into master. DAOD_PHYS is adjusted to uncomment some lines that were broken by this issue and are now able to work.