Fix muon trigger configuration when running RDOtoRDOtrigger
There is some conflict between RecExCommon_topOptions.py and runHLT_standalone.py, where the former is configuring some tools used in the muon trigger before we can set the appropriate flag in runHLT_standalone.py to configure the tools correctly for the trigger. To avoid this, we can set the flag needed for the trigger before including RecExCommon_topOptions.py. This fixes ATR-21707.