PixelConditionsAlgorithms: Fix initialization hack.
PixelCablingCondAlg::initialize() is recording a conditions object
to be read by SiRegionSeletorTable::initialize().  However, it was
doing this with an unbounded IOV range, which then collides with
any other IOV ranges that may be added later.

Change so that the range for this hack covers only run number 0,
so it won't collide with any real IOV ranges.