AthenaServices: Make stack dumps more robust.
Tweak CoreDumpSvc to be more robust against multiple signals from different
threads.

The CoreDumpSvc signal handler had protection against recursion.
But that could be confused if we get signals in multiple threads,
leading to one thread terminating the program while another is trying
to make a stack dump.

Rework to separate out the cases of recursion from that of multiple threads,
and only allow one thread at a time to be dumping.