athenaHLT: Fix missing child histograms
With atlas-tdaq-software/HLTMPPU!14 we made `SkipFinalizeWorker` the
default as this was the default operations mode at P1 since a long time.
However, in athenaHLT this leads to the child histograms not being saved
to the final ROOT file (ATR-21715). As a workaround set
`SkipFinalizeWorker=0` in athenaHLT until we have a proper fix.