TrigPSC: Save child histograms during stopRun
With atlas-tdaq-software/HLTMPPU!14 we made `SkipFinalizeWorker` the
default. In order for the child histograms to be saved with the regular
THistSvc they need to saved already at the end of the stop transition
because finalize never happens. Call `THistSvc::finalize` for the
workers at the end of `stopRun`. This also works in online mode with the
`TrigMonTHistSvc` as the finalize() method of that service is empty.

If run with `SkipFinalizeWorker=0` this also seems to work as the
THistSvc gracefully handles the case where it is being finalized
twice. However, this may of course change at any time. In that case we
should make this behavior dependent on the value of
`SkipFinalizeWorker=0`.

Relates to ATR-21715.