Allow some track manipulation for the cosmic analyses
In some cases the cosmic tracks in the trigger and offline may use a different convention,
offline tracks always travelling downwards, but the trigger, travelling outwards from
the beamline.

These changes allow the direction of th the trigger tracks to be reversed and so restore
the correspondence between trigger and offline tracks.

In additon, because of the wider d0 and z0 ranges that are possible with cosmics, the
histogram ranges can be modified from  dat file.

In addition include changes to allow the xAOD::TruthParticles to be retrieved for the
cosmic MC.