Add new ATLAS 'blue' logos and fix the selection in the Editor.
The 'Blue' selector for the blue version of the ATLAS logo was disabled
in the masterc version of VP1. That because we had no blue versions of
the new logo with transparent background. Thus, I created them from .ai
files (Adobe Illustrator) and saved them in all the needed sizes. Then I
re-enabled the selector and updated the code.