Switch off seeded decoding for full scan Run 2 muon triggers
Somehow the seeded decoding for the full scan muon triggers in the run 2 framework does not always decode the full detector, even though we are decoding an RoI that should cover the full detector. To get around this, will switch off the seeded decoding which by default does decode the full detector.
This matches what is done for the Run 3 trigger, which has the correct behaviour.