Now linking Tauolapp_i against TauolaHepMC.
With the 1.1.8.atlas version of Tauola++, which is part of
LCG_97a_ATLAS_1, the libraries were split up a bit more. And now
we need to link against libTauolaHepMC.so as well.

Note that once the HepMC3 migration reaches this package, I believe
this configuration will have to be updated to use
libTauolaHepMC3.so instead.