AthenaKernel: Extend ThinningCache to hold additional data.
ThinningCache holds a map of object keys to a set of thinned entries,
and is used during output.

Extend ThinningCache so that it now olds a ThinningInfo structure,
rather than just the set of thinned entries.
This allows us to extend ThinningCache to hold additional information.
Here, we're adding the set of vetoed variables per object.