AthenaServices: Use ThinningCache for variable selection.
AthenaOutputStream processes ItemList information, including requests
for dynamic variable information.  To prevent different streams from
interfering with each other, we want to put this information in the
ThinningCache, rather than in the EDM objects themselves.

Changed AthenaOutputStream so that rather than calling selectAux, it records
variable selection information in the event store.  This is then read
by ThinningCacheTool, to make the information available in the ThinningCache