Reject poor quality EM-only tracks
EM segments that are not used in any other tracks are made into MS tracks: however, these tracks are often of quite poor quality, and cannot be extrapolated successfully.
If they cannot be extrapolated, they are essentially never used to form muons, and anyway MuGirl should catch any tracks that MuidCo misses because of this.
So we will now not make MuonCandidates from EM-only MS tracks that cannot be extrapolated.
This will also have the helpful side effect of greatly reducing the number of warnings generated, such as those in ATLASRECTS-4464 and ATLASRECTS-5426.