Reject poor muon track extrapolations
Add a check to reject extrapolated tracks if the chi2 of the new track is much worse than the chi2 of the original MS track.
For now, "much worse" means "1000 times worse": this may have to be adjusted, but the goal is to remove really terrible fits, not just fits that aren't quite as good.
This should reduce the number of warnings of the type of those seen in ATLASRECTS-5427, by ensuring that the poorly fitted track is not used in the combined fit procedure.
Instead, the original MS track will be used, so any loss of muons should be negligible.