Add more muon triggers to LS2 menu
Adding muon triggers to LS2 menu to bring in line with the physics_pp_v7_primaries menu, and included some additional primaries listed in ATR-20049.Fixed a small bug in the invariant mass hypo configuration to pick up the invariant mass from the correct leg of asymmetric chains. Updating the references to include the added chains.