Remove setting of CoreDumpSvc.FatalHandler
With athenaMT it is best to rely on the default (ROOT) signal handler in
CoreDumpSvc, which uses gdb to produce the stacktrace for all threads.
Setting `FatalHandler = 438` makes use of the `SealSignal` handler,
which only produces one trace. For production environments where the gdb
traces are not needed, one should override the setting in the transform
or AMI tag as is already done today.

Relates to ATEAM-629.