Store cluster splitting probabilities in map rather than updating the cluster.
Originally the cluster splitting probabilities were stored in a reserved word of the
corresponding PixelClusters. This however required to update a collection already stored
in StoreGate i.e. was not thread-safe.

Now the splitting probabilities are stored in separate maps, which are copied and updated
when passed from algorithm to algorithm. The algorithms record the map at the very beginning
to allow their tools (PixelClusterOnTrackTool) to access the map via read handles rather than
passing the map down the call chain explicitly.