Migrate any remaining hole search config files
There are a few remaining configuration scripts that do not get used
by q431 but which we should still migrate to use the new boundary
checking tool. This commit adapts those files to use the new tool,
often creating a new tool for it and using that in lieu of the old
configuration parameters.