Another fix for ATLASRECTS-4909
In some extremely rare cases it may be impossible to obtain a caloExtension for a combined track.
This seems to happen only for very low-pT muons, for which the extrapolation is not robust, so that starting from the perigee or the edge of the ID can determine whether it succeeds.
For technical reasons, during the combined fit the caloExtension is usually determined from the perigee, while the caloIsolationTool will start from the edge of the ID.
Therefore it is possible for a caloExtension to be successfully obtained during the combined fit but not later.
This fix will prevent muons being built from combined tracks in the case that a caloExtension cannot be obtained.