TriggerMenu: Fix irreproducibilty.
Some of the jet menu code was using the builtin hash() function
to form trigger element names.  With py3, however, the result of hash()
can vary from run to run.  This implies that the HLTResult object
written to AOD files can also change from run to run.

Change the use of hash() to crc32() in order to make this
reproducible.