Add muon candidate and muon creator algs to trigger new JO test
Adding the building of muoncandidates and xAOD muons for the SA part of the muon trigger in the new JO test. In principle this lets us run ms-only type chains in the new configuration (or will once I add a sensible hypo in the future).
- Made some changes in the configuration to various tools for trigger running
- Added configuration for alg to convert MS tracks to xADO trackParticles (temporarily in trigger config)
- Updated trigger configuration to call muon candidate and creator algs

Part of the migration in ATR-20388