AthenaCommon: Remove setting of CoreDumpSvc.FatalHandler
Follow-up to !35076 where we removed the explicit setting of
`CoreDumpSvc.FatalHandler` in various job options. However, `athena.py`
is setting this value to `-1` (only handle `SIGINT`) by default and this
leads to missing stacktraces for real crashes (e.g. ATR-21816). Not
clear why this setting was chosen as a default in athena.

Also update boostrap pickle files to match new default.