Decide whether SimHit Collections should be created per simulator on a per region-basis
In ATLFASTII we only use two simulators (G4 and FCS V1) in the Calorimeters, it does not
make sense to create simulator_specific collections and then "merge" them (effectively a straight copy)
for other the sub-systems. This change allows the CollectionMerger algorithm to be configured separately for
ID, CALO and MUON regions of the detector. This allows us to switch off merging in the ID and MUON regions
in ATLFASTII and G4FastCalo jobs. In this initial version the configuration of other top level configurations is unchanged.