AthenaKernel: Use static_cast to DataBucket.
A DataProxy holds a pointer to a DataObject.
When we were doing a casting operation from a DataProxy,
we were using dynamic_cast to convert from DataObject to DataBucketBase.
Nowadays, though, this DataObject is always an instance of a DataBucket.
So do a static_cast instead.
This speeds up the casting operation.