xAODTrigRingerCnv: Delete package
While this package contains code that is being compiled into a component
library it misses the `xAODTrigRingerCnv_entries.cxx` file and hence
none of these components could have ever been used in athena. There are
also no relevant code changes in many years.

The best guess is that this package was created during AOD-times but
when the Ringer algorithm was first deployed we were already in the
xAOD-era and so this converter code was never needed.

Part of trigger package cleanup ATR-19150.