Removed Trig::WebBunchCrossingTool from TrigBunchCrossingTool.
As discussed in ATR-21827, atlas-trigconf.cern.ch now moved behind
CERN SSO. Meaning that this tool is not usable anymore.

If we change our minds about this later on, we will still be able
to revive the code from Git's history.

At the same time brought in some updates to one of the test executables
from 21.2. But will use a different commit for enabling that test's
build in master.