Fixes to MDT_DigitizationConfig module
Updated `MDT_Response_DigiToolCfg` and `RT_Relation_DB_DigiToolCfg` to return a
`ComponentAccumulator` object. Better than returning a bare tool as then clients
do not need to be updated if additional Tool/ServiceHandles are added to these tools.

Updated `MDT_DigitizationToolCommonCfg` to explicitly use `MdtCalibrationDbToolCfg`
in the case that the `GetT0FromBD` property is True. (This is only the case for data
overlay jobs currently.) In this way `MdtCalibDbAlg` is added to the `ComponentAccumulator`
when required as part of the `MdtCalibrationDbToolCfg` call. Also updated
`RT_Relation_DB_DigiToolCfg` to use `MdtCalibrationDbToolCfg`, although currently this method
is not used in standard production. It should be noted that in the current version of
the code using `RT_Relation_DB_DigiTool` would require Muon alignments to be switched on during
digitization.