Refactoring of MuonGeometryConfig module
Create separate configuration methods for `MuonAlignmentCondAlgCfg`, `MuonDetectorToolCfg`
and `MuonDetectorCondAlgCfg`. This makes it clearer which components are using
which conditions folders and makes the `MuonGeoModelCfg` method shorter. There are some
repeated `if ... else` blocks in `MuonAlignmentCondAlgCfg` and `MuonDetectorToolCfg` where
properties are set depending on the same `ConfigFlags`.

`MdtCalibDbAlg` uses the `MuonDetectorManager`, which is provided by `MuonDetectorCondAlg`,
so added an explicit call to `MuonDetectorCondAlg` to the `MdtCalibDbAlgCfg` method.