Patching EXOT4 to ensure truth is available when needed
There was a minor ordering issue in EXOT4, where truth algorithms were
being added to the derivation after the main sequence, which needed some
of those algorithms. This MR just swaps the order, and thus fixes the
derivation. There's quite a bit of other clean up that would be possible
in this derivation code, but I'll leave that to the contact.