Deduplicate track parameter calculation propagator
In the existing code base, the GlobalChi2Fitter performs propagations
in order to update its track parameters. It does this for upstream as
well as downstream states, and it can try this in up to four different
branches depending on a set of conditions. This leads to massive
amounts of code duplication. That makes the code harder to debug, more
prone to errors, and harder to maintain and change in the future.
Since our goal is to use adjust this code to do hole search, the
current state of it is unworkable.

This commit lifts the propagation code out of the track parameter
updating code and into two new helper methods which deduplicate the
method. This reduces code volume and increases clarity. Changes are
also made to the memory ownership model of the code, switching to smart
pointers wherever possible.